Completed Parser to route new sql init file to db/init/003_seed.sql (run 'python db_loader.py' in /src to generate). Completed subject_map.py to map subject codes to names. Running into problems loading records into database, requires set order of precedence when inserting tables maybe???

+++ db/init/001_schema.sql
@@ -93,7 +93,7 @@ CREATE INDEX idx_section_instructor_id ON section(instructor_id);
 -- =========================
 CREATE TABLE grade_distribution (
     section_id        INT PRIMARY KEY,       -- PK & FK, NOT NULL
-    gpa               INT NOT NULL,          -- overall GPA for this section
+    gpa               DECIMAL(10,2) NOT NULL,          -- overall GPA for this section
 
     a                 INT NOT NULL DEFAULT 0,
     a_minus           INT NOT NULL DEFAULT 0,
